\name{addsuffixes}
\alias{addsuffixes}
\title{addsuffixes}
\usage{addsuffixes(names=c("T0001", "T0099", "T0001", "T0099"),
    suffix=c("_Tx", "_Tn"), sep="")}
\description{Adds suffixes for daily maximum and minimum temperature to the names of a column data frame}
\author{Emanuele Cordano, Emanuele Eccel}
\seealso{\code{\link{getVARmodel}}}
\value{the vector of names with suffixes added}
\arguments{\item{names}{a character string vector with column names}
\item{suffix}{suffixes to add to the first and second groups of column names respectively}
\item{sep}{separation element}}
