\name{arch_test}
\alias{arch_test}
\title{\code{arch.test} function for \code{varest2} object}
\usage{
  arch_test(object, ...)
}
\arguments{
  \item{object}{a \code{varest2} object}

  \item{...}{passed arguments}
}
\description{
  \code{arch.test} function for \code{varest2} object
}
\seealso{
  \code{\link{arch.test}}
}

