\name{collinear_dataset}
\alias{collinear_dataset}
\docType{data}


\title{
Collinear Dataset
}
\usage{
data(collinear_dataset)
}

\description{
It is an artificial example dataset contaning 16 variables with collinearity among some of them.
}
\details{
The user can easily use the package with his/her own data after replacing the values of such variables.
}
\format{
Data frame
}
\source{


This dataset is intended for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.
}

%%\references{
%% ~put references to the literature/web site here ~
%%}
\keyword{dataset}
