\name{countNAs}
\alias{countNAs}
\title{counts NAs in each row of \code{data}}
\usage{
  countNAs(data)
}
\arguments{
  \item{data}{a data input matrix}
}
\value{
  the vector of names with added suffixes
}
\description{
  counts NAs in each row of \code{data}
}
\author{
  Emanuele Cordano, Emanuele Eccel
}

