% Generated by roxygen2 (4.0.2): do not edit by hand
\name{forecastEV}
\alias{forecastEV}
\title{Forecasts the expected value of a VAR realization given the prievious one}
\usage{
forecastEV(var, xprev = NULL, exogen = NULL)
}
\arguments{
\item{var}{A VAR model represented by a \code{varest} object as returned by \code{\link{getVARmodel}} or \code{\link{VAR}}}

\item{xprev}{previous status of the random variable}

\item{exogen}{vector containing the values of the "exogen" variables (predictor) for the generation}
}
\value{
a vector of values
}
\description{
Forecasts the expected value of a VAR realization given the prievious one
}
\author{
Emanuele Cordano, Emanuele Eccel
}
\seealso{
\code{\link{forecastResidual}}
}

