% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDailyMean.R
\name{getDailyMean}
\alias{getDailyMean}
\title{Calculates the daily means of a range of days around each date of a data frame corresponding to a period between \code{year_min} and \code{year_max}  for stations listed in \code{station}}
\usage{
getDailyMean(
  data,
  year_min = 1961,
  year_max = 1990,
  station = c("T0001", "T0010"),
  origin = "1961-1-1",
  lag = 5
)
}
\arguments{
\item{data}{a data frame containing daily data.}

\item{year_min}{start year}

\item{year_max}{end year}

\item{station}{character vector of the IDs of the station where the data are requested}

\item{origin}{origin date of time-series}

\item{lag}{lag (number of days) on which daily mean is calculated. The mean is calculated considereing \code{lag} days before and after each day.}
}
\value{
a matrix containing the requested daily mean data where each day corresponds to a row and each station corresponds to a column
}
\description{
Calculates the daily means of a range of days around each date of a data frame corresponding to a period between \code{year_min} and \code{year_max}  for stations listed in \code{station}
}
\note{
The input data frame \code{data} must have the following fields: \code{year,month,day,variables_ID1,variables_ID2,...} 
where the fields \code{,variables_ID1,variables_ID2,...} contain the daily variables referred to the respective stations and the field names are replaced with the respective station ID.
}
\seealso{
\code{\link{extractyears}}
}
\author{
Emanuele Cordano, Emanuele Eccel
}
