% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zc0_GPCA.R
\docType{class}
\name{GPCA-class}
\alias{GPCA-class}
\title{GPCA-class}
\description{
\code{GPCA} S3 class returned by \code{\link{GPCA}}
}
\details{
\describe{
	\item{list of \code{GPCA_iteration}}{subsequent GPCA iterations}

\item{\code{final_results} }{data.frame or matrix of the "gaussianized" data}
 
}
}
\note{
Formal definition with \code{\link{setOldClass}} for the S3 class \code{GPCA}
}
\examples{
showClass("GPCA")

 

}
\author{
Emanuele Cordano
}
\keyword{classes}

