% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normality_test}
\alias{normality_test}
\title{\code{normality.test} method for \code{varest2} object}
\usage{
normality_test(object, ...)
}
\arguments{
\item{object}{a \code{varest2} object}

\item{...}{passed arguments}
}
\description{
\code{normality.test} method for \code{varest2} object
}
\seealso{
\code{\link{normality.test}}
}

