% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeGaussian_sevaralstations.R
\name{normalizeGaussian_severalstations}
\alias{normalizeGaussian_severalstations}
\title{Converts several samples \code{x} random variable  extracted by populations represented by the columns of \code{data} respectively or \code{sample}
 to a normally-distributed samples with assinged mean and standard deviation or vice versa in case \code{inverse} is \code{TRUE}}
\usage{
normalizeGaussian_severalstations(x, data = x, cpf = NULL, mean = 0,
  sd = 1, inverse = FALSE, step = NULL, prec = 10^-4, type = 3,
  extremes = TRUE, sample = NULL, origin_x = NULL,
  origin_data = NULL)
}
\arguments{
\item{x}{value to be converted}

\item{data}{a sample of data on which a non-parametric probability distribution is estimated}

\item{cpf}{cumulative probability distribution. If \code{NULL} (default) is calculated as \code{\link{ecdf}(data)}}

\item{mean}{mean (expected value) of the normalized random variable. Default is 0.}

\item{sd}{standard deviation of the normalized random variable. Default is 1.}

\item{inverse}{logical value. If \code{TRUE} the function works inversely (the opposite way). Default is \code{FALSE}.}

\item{step}{vector of values in which step discontinuities of the cumulative probability function occur. Default is \code{NULL}}

\item{prec}{amplitude of the neighbourhood of the step discontinuities where cumulative probability function is treated as non-continuous.}

\item{type}{see \code{\link{quantile}}}

\item{extremes}{logical variable. 
If \code{TRUE} (default) the probability or frequency is multiplied by \deqn{\frac{N}{N+1}} where \eqn{N} is the length of \code{data}}

\item{sample}{information on how to sample \code{x} and \code{data}. Default is \code{NULL}, this means that the values of each column of \code{x} and \code{data} belong to the same sample. If \code{x} and \code{data} are sampled for each month seperately, it is set to \code{monthly}.}

\item{origin_x}{date corresponding to the first row of \code{x}}

\item{origin_data}{date corresponding to the first row of \code{data}}
}
\value{
a matrix with the normalized variable or its inverse
}
\description{
Converts several samples \code{x} random variable  extracted by populations represented by the columns of \code{data} respectively or \code{sample}
 to a normally-distributed samples with assinged mean and standard deviation or vice versa in case \code{inverse} is \code{TRUE}
}
\note{
It applies \code{\link{normalizeGaussian}} for each column of \code{x} and \code{data}.
See the R code for further details
}
\examples{

library(RMAWGEN) 
N <- 30
x <- rexp(N)
y <- x+rnorm(N)
df <- data.frame(x=x,y=y)

dfg <- normalizeGaussian_severalstations(df,data=df,extremes=TRUE,inverse=FALSE)

dfi <- normalizeGaussian_severalstations(dfg,data=df,extremes=TRUE,inverse=TRUE)


 

}
\seealso{
\code{\link{normalizeGaussian}}
}
\author{
Emanuele Cordano, Emanuele Eccel
}
