\name{print.GPCA}
\alias{print.GPCA}
\title{\code{print} S3 method for \code{GPCA} object}
\usage{
  print.GPCA(x, rmin = 1, rmax = 4, cmin = rmin,
    cmax = rmax, ...)
}
\arguments{
  \item{x}{a \code{GPCA} object}

  \item{rmin,rmax,cmin,cmax}{maximum and minimum rows and
  columns to be printed}

  \item{...}{passed arguments}
}
\description{
  \code{print} S3 method for \code{GPCA} object
}
\seealso{
  \code{\link{GPCA}}
}

