\name{print.GPCAiteration}
\alias{print.GPCAiteration}
\title{\code{print} S3 method for \code{GPCAiteretion} object}
\usage{
  print.GPCAiteration(x, rmin = 1, rmax = 4, cmin = rmin,
    cmax = rmax, ...)
}
\arguments{
  \item{x}{a \code{GPCAiteration} object}

  \item{rmin,rmax,cmin,cmax}{maximum and minimum rows and
  columns to be printed}

  \item{...}{passed arguments}
}
\description{
  \code{print} S3 method for \code{GPCAiteretion} object
}
\seealso{
  \code{\link{GPCA_iteration}}
}

