% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.GPCA.R, R/print.GPCAiteration.R
\name{print.GPCA}
\alias{print}
\alias{print.GPCA}
\alias{print.GPCAiteration}
\title{\code{print} S3 method for \code{GPCA} or \code{GPCA_iteration} object}
\usage{
\method{print}{GPCA}(x, rmin = 1, rmax = 4, cmin = rmin, cmax = rmax,
  ...)

\method{print}{GPCAiteration}(x, rmin = 1, rmax = 4, cmin = rmin,
  cmax = rmax, ...)
}
\arguments{
\item{x}{a \code{GPCA} or \code{GPCAiteration} object}

\item{rmin, rmax, cmin, cmax}{maximum and minimum rows and columns to be printed}

\item{...}{passed arguments}
}
\description{
\code{print} S3 method for \code{GPCA} or \code{GPCA_iteration} object
}
\seealso{
\code{\link{GPCA}},\code{\link{GPCA_iteration}}

\code{\link{GPCA_iteration}}
}

