% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaling_monthly.R
\name{rescaling_monthly}
\alias{rescaling_monthly}
\title{This function adjusts the monthly mean to a daily weather dataset (e. g. spline-interpolated temperature)}
\usage{
rescaling_monthly(data, val, origin = "1961-1-1")
}
\arguments{
\item{data}{data frame of wheather variables)}

\item{val}{monthly means returned by \code{\link{getMonthlyMean}}}

\item{origin}{character string containing the gregorian date of the first day of \code{data}}
}
\value{
A data frame with data of \code{data} rescaled with \code{val} for each month
}
\description{
This function adjusts the monthly mean to a daily weather dataset (e. g. spline-interpolated temperature)
}
\note{
It uses \code{\link{months}} and  \code{\link{julian}}
}
\author{
Emanuele Cordano


 @export
}
\seealso{
\code{\link{extractdays}}
}

