\name{residuals.varest2}
\alias{residuals}
\alias{residuals.varest2}
\title{\code{residuals} S3 method for \code{varest2} object}
\usage{
  \method{residuals}{varest2} (object, ...)
}
\arguments{
  \item{object}{a \code{blockmatrix} object}

  \item{...}{passed arguments}
}
\value{
  residuals of blockmatrix \code{object}
}
\description{
  \code{residuals} S3 method for \code{varest2} object
}
\author{
  Emanuele Cordano
}

