% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial_test.varest2.R
\name{serial_test}
\alias{serial_test}
\title{\code{serial.test} function for \code{varest2} object}
\usage{
serial_test(object, ...)
}
\arguments{
\item{object}{a \code{varest2} object}

\item{...}{passed arguments}
}
\description{
\code{serial.test} function for \code{varest2} object
}
\seealso{
\code{\link{serial.test}}
}

