% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemperatureEndDay.R
\name{TemperatureEndDay}
\alias{TemperatureEndDay}
\title{Gets the last day in a temperature time series, expressed as decimal julian days since 1970-1-1 00:00 UTC}
\usage{
TemperatureEndDay(name, station_names, end_day)
}
\arguments{
\item{name}{character ID of the station}

\item{station_names}{vector containing the IDs (characters)  of the considered meteorological stations. An example is \code{STATION_NAMES} defined in the \code{\link{trentino}} dataset.}

\item{end_day}{vector containing the measurement end day. An example is \code{TEMPERATURE_MEASUREMENT_END_DAY} defined in the \code{\link{trentino}} dataset.}
}
\value{
the temperature measurement end day given the vectors of station IDs and the temperature measurement end days
}
\description{
Gets the last day in a temperature time series, expressed as decimal julian days since 1970-1-1 00:00 UTC
}
\examples{
data(trentino)
TemperatureEndDay("T0099",station_names=STATION_NAMES,end_day=TEMPERATURE_MEASUREMENT_END_DAY) 
}
\author{
Emanuele Cordano, Emanuele Eccel
}

