% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zc0_varest.R
\docType{class}
\name{varest-class}
\alias{varest-class}
\alias{varest}
\title{varest-class}
\description{
\code{varest} S3 class (formal definition) see \code{\link{VAR}}
}
\details{
\describe{
    The details of the class are reported on \code{\link{VAR}} documentation in "vars" package
 }
}
\note{
Formal definition with \code{\link{setOldClass}} for the S3 class \code{varest}
}
\examples{
showClass("varest")

 

}
\author{
Bernhard Pfaff
}
\keyword{classes}
