\docType{class}
\name{varest2-class}
\alias{varest2}
\alias{varest2-class}
\title{varest2-class}
\description{
This class derives from a \code{varest} S3 class which is a
list of objects describing a Vectorial AutoRegressive Model
(see \code{\link{VAR}})
}
\details{
\describe{ \item{\code{VAR}:}{a \code{varest} S3 object
created by \code{\link{VAR}} } }
}
\note{
A \code{varest2} object can be created by
\code{new("varest2", ...)} or returned by the function
\code{\link{getVARmodel}}
}
\examples{
showClass("varest2")
}
\author{
Emanuele Cordano
}
\keyword{classes}

