% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adddate}
\alias{adddate}
\title{Inserts three columns (year,month,day) passing dates  to a matrix or to a dataframe}
\usage{
adddate(data, origin = "1961-1-1")
}
\arguments{
\item{data}{matrix of daily data}

\item{origin}{character string containing the date of the first row of \code{data} as YYYY-MM-DD}
}
\value{
a data frame with dates and \code{data} values
}
\description{
Inserts three columns (year,month,day) passing dates  to a matrix or to a dataframe
}
\author{
Emanuele Cordano, Emanuele Eccel
}
\seealso{
\code{\link{findDate}}
}

