% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addsuffixes.R
\name{addsuffixes}
\alias{addsuffixes}
\title{Adds suffixes for daily maximum and minimum temperature to the names of a column data frame}
\usage{
addsuffixes(names = c("T0001", "T0099", "T0001", "T0099"),
  suffix = c("_Tx", "_Tn"), sep = "")
}
\arguments{
\item{names}{a character string vector with column names}

\item{suffix}{suffixes to add to the first and second groups of column names respectively}

\item{sep}{separation element}
}
\value{
the vector of names with suffixes added
}
\description{
Adds suffixes for daily maximum and minimum temperature to the names of a column data frame
}
\details{
This function is used for data frames with duplicated field names
}
\examples{
names <- addsuffixes()
}
\seealso{
\code{\link{getVARmodel}}
}
