% Generated by roxygen2 (4.0.2): do not edit by hand
\name{countNAs}
\alias{countNAs}
\title{counts NAs in each row of \code{data}}
\usage{
countNAs(data)
}
\arguments{
\item{data}{a data input matrix}
}
\value{
the vector with numbers of NA values for each \code{data} column
}
\description{
counts NAs in each row of \code{data}
}
\author{
Emanuele Cordano, Emanuele Eccel
}

