% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNAs.R
\name{countNAs}
\alias{countNAs}
\title{counts NAs in each row of \code{data}}
\usage{
countNAs(data)
}
\arguments{
\item{data}{a data input matrix

@export}
}
\value{
the vector with numbers of NA values for each \code{data} column
}
\description{
counts NAs in each row of \code{data}
}
