% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qqplotWGEN}
\alias{qqplotWGEN}
\title{Makes a qqplot and Wilcoxon test between the two columns of \code{val}}
\usage{
qqplotWGEN(val, xlab = "simulated", ylab = "measured", main = "title",
  ylim = c(min(val), max(val)), xlim = c(min(val), max(val)),
  diff = FALSE, quantile = 0)
}
\arguments{
\item{val}{a matrix with two columns containing the two samples to be compared}

\item{xlab,ylab,main}{see \code{\link{plot.default}}}

\item{xlim,ylim}{see \code{\link{plot.default}}}

\item{diff}{logical variable, if \code{TRUE} the function is applied to \code{diff(val)} instead of \code{val}. See \code{\link{diff}}}

\item{quantile}{quantile value on which data samples in \code{val} are considered. Default is 0.}
}
\value{
Wilcoxon test between the two columns of 'val'
}
\description{
Makes a qqplot and Wilcoxon test between the two columns of \code{val}
}
\author{
Emanuele Cordano, Emanuele Eccel
}

