% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplotWGEN.R
\name{qqplotWGEN}
\alias{qqplotWGEN}
\title{Makes a qqplot and Wilcoxon test between the two columns of \code{val}}
\usage{
qqplotWGEN(val, xlab = "simulated", ylab = "measured", main = "title",
  ylim = c(min(val), max(val)), xlim = c(min(val), max(val)),
  diff = FALSE, quantile = 0)
}
\value{
Wilcoxon test between the two columns of 'val'
}
\description{
@param val a matrix with two columns containing the two samples to be compared
 @param xlab,ylab,main see \code{\link{plot.default}}
 @param xlim,ylim see \code{\link{plot.default}}
 @param diff logical variable, if \code{TRUE} the function is applied to \code{diff(val)} instead of \code{val}. See \code{\link{diff}}
	@param quantile quantile value on which data samples in \code{val} are considered. Default is 0.
}
\author{
Emanuele Cordano, Emanuele Eccel
}

