% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNAs.R
\name{removeNAs}
\alias{removeNAs}
\title{Replaces each entry of the rows containing NA values with NA}
\usage{
removeNAs(data)
}
\arguments{
\item{data}{a matrix 


 @author  Emanuele Cordano, Emanuele Eccel}
}
\value{
the matrix \code{data} with the modified rows of NA values
}
\description{
Replaces each entry of the rows containing NA values with NA
}
\note{
In \code{\link{getVARmodel}}, 
when using \code{\link{VAR}} or \code{\link{VARselect}}, all NAs will be removed
}
\seealso{
\code{\link{getVARmodel}}
}

