\docType{package}
\name{RMAWGEN-package}
\alias{RMAWGEN}
\alias{RMAWGEN-package}
\title{R - Multi-site Autoregressive WEather Generator}
\description{
  Multi-site autoregressive Models for Daily Weather
  Generation
}
\details{
  \tabular{ll}{ Package: \tab RMAWGEN\cr Type: \tab
  Package\cr Version: \tab 1.2\cr Date: \tab 2011-11-22\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr
  Depends: R(>=2.12),time,chron,vars \cr }
}
\author{
  Emanuele Cordano \email{emanuele.cordano@gmail.org},
  Emanuele Eccel \email{emanuele.eccel@iasma.it}
}
\keyword{package}

