% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/splineInterpolateMonthlytoDailyforSeveralYears.R
\name{splineInterpolateMonthlytoDailyforSeveralYears}
\alias{splineInterpolateMonthlytoDailyforSeveralYears}
\title{Interpolates monthly data to daily data using \code{\link{splineInterpolateMonthlytoDaily}}  for several years}
\usage{
splineInterpolateMonthlytoDailyforSeveralYears(val, start_year = 2010,
  nyear = 1, leap = TRUE, offset = 2, no_spline = FALSE,
  yearly = FALSE)
}
\arguments{
\item{val}{matrix containing monthly mean data for one year}

\item{start_year}{first year}

\item{nyear}{number of years since \code{start_year}}

\item{leap}{logical variable If \code{TRUE} (default) leap years are considered, otherwise they are not}

\item{offset}{integer values. Default is 2. Number of years considered beyond the extremes in order to avoid edge errors}

\item{no_spline}{logical value. If \code{TRUE} no spline interpolation is calculated and the daily value corresponds to the monthly average value. Default is \code{FALSE}.}

\item{yearly}{logical value. If \code{TRUE} the result with men value per each month per each year. Default is \code{FALSE}.


 @return a matrix or data frame with interpolated daily data}
}
\description{
Interpolates monthly data to daily data using \code{\link{splineInterpolateMonthlytoDaily}}  for several years
}
\seealso{
\code{\link{spline}},\code{\link{splineInterpolateMonthlytoDaily}}
}
\author{
Emanuele Cordano, Emanuele Eccel
}
